+++ buildSrc/src/main/kotlin/io/getstream/chat/android/Configuration.kt
@@ -6,8 +6,8 @@ object Configuration {
     const val sampleTargetSdk = 33
     const val minSdk = 21
     const val majorVersion = 5
-    const val minorVersion = 15
-    const val patchVersion = 0
+    const val minorVersion = 17
+    const val patchVersion = 4
     const val versionName = "$majorVersion.$minorVersion.$patchVersion"
     const val snapshotVersionName = "$majorVersion.$minorVersion.${patchVersion + 1}-SNAPSHOT"
     const val artifactGroup = "io.getstream"