+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/factory/FileAttachmentFactory.kt
@@ -19,18 +19,28 @@ package io.getstream.chat.android.compose.ui.attachments.factory
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.runtime.Composable
+import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.content.FileAttachmentContent
 import io.getstream.chat.android.compose.ui.attachments.content.FileAttachmentPreviewContent
+import io.getstream.chat.android.compose.ui.attachments.content.onFileAttachmentContentItemClick
+import io.getstream.chat.android.compose.ui.attachments.preview.handler.AttachmentPreviewHandler
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.uiutils.extension.isFile
 
 /**
  * An [AttachmentFactory] that validates attachments as files and uses [FileAttachmentContent] to
  * build the UI for the message.
+ *
+ * @param onContentItemClick Lambda called when an item gets clicked.
  */
 @Suppress("FunctionName")
-public fun FileAttachmentFactory(): AttachmentFactory = AttachmentFactory(
+public fun FileAttachmentFactory(
+    onContentItemClick: (
+        previewHandlers: List<AttachmentPreviewHandler>,
+        attachment: Attachment,
+    ) -> Unit = ::onFileAttachmentContentItemClick,
+): AttachmentFactory = AttachmentFactory(
     canHandle = { attachments ->
         attachments.any { it.isFile() }
     },
@@ -46,7 +56,8 @@ public fun FileAttachmentFactory(): AttachmentFactory = AttachmentFactory(
             modifier = modifier
                 .wrapContentHeight()
                 .width(ChatTheme.dimens.attachmentsContentFileWidth),
-            attachmentState = state
+            attachmentState = state,
+            onItemClick = onContentItemClick,
         )
     },
 )