+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/composer/MessageComposerViewStyle.kt
@@ -64,9 +64,12 @@ import io.getstream.chat.android.ui.utils.extensions.getDrawableCompat
  * @param messageInputMaxLines The maximum number of message input lines.
  * @param messageInputCannotSendHintText The input hint text in case we can't send messages in this channel.
  * @param messageInputInputType The [InputType] to be applied to the message input edit text.
+ * @param messageInputShowReplyView Whether to show the default reply view inside the message input or not.
  * @param attachmentsButtonVisible If the button to pick attachments is displayed.
  * @param attachmentsButtonIconDrawable The icon for the attachments button.
  * @param attachmentsButtonRippleColor Ripple color of the attachments button.
+ * @param takeImageEnabled If starting image capture is enabled.
+ * @param recordVideoEnabled If starting video capture is enabled.
  * @param commandsButtonVisible If the button to select commands is displayed.
  * @param commandsButtonIconDrawable The icon for the commands button.
  * @param commandsButtonRippleColor Ripple color of the commands button.
@@ -122,10 +125,13 @@ public data class MessageComposerViewStyle(
     public val messageInputMaxLines: Int,
     public val messageInputCannotSendHintText: String,
     public val messageInputInputType: Int,
+    public val messageInputShowReplyView: Boolean,
     // Leading content
     public val attachmentsButtonVisible: Boolean,
     public val attachmentsButtonIconDrawable: Drawable,
     @ColorInt public val attachmentsButtonRippleColor: Int?,
+    public val takeImageEnabled: Boolean,
+    public val recordVideoEnabled: Boolean,
     public val commandsButtonVisible: Boolean,
     public val commandsButtonIconDrawable: Drawable,
     @ColorInt public val commandsButtonRippleColor: Int?,
@@ -400,6 +406,14 @@ public data class MessageComposerViewStyle(
                     R.styleable.MessageComposerView_streamUiMessageComposerAttachmentsButtonVisible,
                     true
                 )
+                val takeImageEnabled = a.getBoolean(
+                    R.styleable.MessageComposerView_streamUiMessageComposerTakeImageEnabled,
+                    true
+                )
+                val recordVideoEnabled = a.getBoolean(
+                    R.styleable.MessageComposerView_streamUiMessageComposerRecordVideoEnabled,
+                    true
+                )
 
                 val attachmentsButtonIconDrawable = a.getDrawableCompat(
                     context,
@@ -524,6 +538,11 @@ public data class MessageComposerViewStyle(
                         InputType.TYPE_TEXT_FLAG_CAP_SENTENCES
                 )
 
+                val messageInputShowReplyView = a.getBoolean(
+                    R.styleable.MessageComposerView_streamUiMessageComposerShowMessageReplyView,
+                    true
+                )
+
                 val mediumTypeface = ResourcesCompat.getFont(context, R.font.stream_roboto_medium) ?: Typeface.DEFAULT
 
                 val messageReplyBackgroundColor: Int =
@@ -625,8 +644,11 @@ public data class MessageComposerViewStyle(
                     messageInputMaxLines = messageInputMaxLines,
                     messageInputCannotSendHintText = messageInputCannotSendHintText,
                     messageInputInputType = messageInputInputType,
+                    messageInputShowReplyView = messageInputShowReplyView,
                     // Leading content
                     attachmentsButtonVisible = attachmentsButtonVisible,
+                    takeImageEnabled = takeImageEnabled,
+                    recordVideoEnabled = recordVideoEnabled,
                     attachmentsButtonIconDrawable = attachmentsButtonIconDrawable,
                     attachmentsButtonRippleColor = attachmentsButtonRippleColor,
                     commandsButtonVisible = commandsButtonVisible,
@@ -655,7 +677,6 @@ public data class MessageComposerViewStyle(
                     messageReplyTextStyleTheirs = messageReplyTextStyleTheirs,
                     messageReplyMessageBackgroundStrokeColorTheirs = messageReplyMessageBackgroundStrokeColorTheirs,
                     messageReplyMessageBackgroundStrokeWidthTheirs = messageReplyMessageBackgroundStrokeWidthTheirs,
-
                 ).let(TransformStyle.messageComposerStyleTransformer::transform)
             }
         }